Collapse mediatr down to command and query directories
