+++ src/todo.application/Commands/AddTaskCommand.cs
@@ -5,7 +5,7 @@ using Alteridem.Todo.Domain.Entities;
 using Alteridem.Todo.Domain.Interfaces;
 using MediatR;
 
-namespace Alteridem.Todo.Application.Commands.Add
+namespace Alteridem.Todo.Application.Commands
 {
     public sealed class AddTaskCommand : IRequest<TaskItem>
     {
