+++ src/todo.application/Commands/ArchiveTasksCommand.cs
@@ -6,7 +6,7 @@ using Alteridem.Todo.Domain.Entities;
 using Alteridem.Todo.Domain.Interfaces;
 using MediatR;
 
-namespace Alteridem.Todo.Application.Commands.Archive
+namespace Alteridem.Todo.Application.Commands
 {
     public sealed class ArchiveTasksCommand : IRequest<IList<TaskItem>>
     {
