+++ src/todo.application/Commands/DeleteTaskCommand.cs
@@ -7,7 +7,7 @@ using Alteridem.Todo.Domain.Entities;
 using Alteridem.Todo.Domain.Interfaces;
 using MediatR;
 
-namespace Alteridem.Todo.Application.Commands.Delete
+namespace Alteridem.Todo.Application.Commands
 {
     public class DeleteTaskCommand : IRequest<int>
     {
