+++ src/todo.application/Commands/DeleteTermCommand.cs
@@ -6,7 +6,7 @@ using Alteridem.Todo.Domain.Entities;
 using Alteridem.Todo.Domain.Interfaces;
 using MediatR;
 
-namespace Alteridem.Todo.Application.Commands.Delete
+namespace Alteridem.Todo.Application.Commands
 {
     public class DeleteTermCommand : IRequest<DeleteTermResult>
     {
