+++ src/todo.application/Commands/PrependCommand.cs
@@ -5,7 +5,7 @@ using Alteridem.Todo.Domain.Entities;
 using Alteridem.Todo.Domain.Interfaces;
 using MediatR;
 
-namespace Alteridem.Todo.Application.Commands.Prepend
+namespace Alteridem.Todo.Application.Commands
 {
     public sealed class PrependCommand : IRequest<TaskItem>
     {
