+++ src/todo.application/Queries/ListAllResponse.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Alteridem.Todo.Domain.Entities;
 
-namespace Alteridem.Todo.Application.Queries.List
+namespace Alteridem.Todo.Application.Queries
 {
     public sealed class ListAllResponse
     {
