+++ src/todo.application/Queries/ListContextsQuery.cs
@@ -7,9 +7,9 @@ using Alteridem.Todo.Domain.Entities;
 using Alteridem.Todo.Domain.Interfaces;
 using MediatR;
 
-namespace Alteridem.Todo.Application.Queries.ListProjects
+namespace Alteridem.Todo.Application.Queries
 {
-    public class ListProjectsQuery : IRequest<string[]>
+    public class ListContextsQuery : IRequest<string[]>
     {
 
         private string[] _terms;
@@ -17,29 +17,29 @@ namespace Alteridem.Todo.Application.Queries.ListProjects
         public string[] Terms { get => _terms ?? new string[0]; set => _terms = value; }
     }
 
-    public sealed class ListProjectsQueryHandler : IRequestHandler<ListProjectsQuery, string[]>
+    public sealed class ListContextsQueryHandler : IRequestHandler<ListContextsQuery, string[]>
     {
         private readonly ITaskFile _taskFile;
         private readonly ITaskConfiguration _config;
 
-        public ListProjectsQueryHandler(ITaskFile taskFile, ITaskConfiguration config)
+        public ListContextsQueryHandler(ITaskFile taskFile, ITaskConfiguration config)
         {
             _taskFile = taskFile;
             _config = config;
         }
 
-        public Task<string[]> Handle(ListProjectsQuery request, CancellationToken cancellationToken)
+        public Task<string[]> Handle(ListContextsQuery request, CancellationToken cancellationToken)
         {
             var tasks = _taskFile.LoadTasks(_config.TodoFile);
             string[] positiveTerms = request.Terms
-                .Where(t => t.StartsWith("+"))
+                .Where(t => t.StartsWith("@"))
                 .ToArray();
             string[] negativeTerms = request.Terms
-                .Where(t => t.StartsWith("-+"))
+                .Where(t => t.StartsWith("-@"))
                 .Select(t => t.Substring(1))
                 .ToArray();
             IEnumerable<string> search = tasks
-                .SelectMany(t => t.ProjectTags)
+                .SelectMany(t => t.ContextTags)
                 .OrderBy(p => p)
                 .Distinct();
             if (positiveTerms.Length > 0)
