+++ src/todo/TodoApplication.cs
@@ -1,25 +1,12 @@
 using System;
 using System.CommandLine;
-using System.CommandLine.Invocation;
 using System.CommandLine.NamingConventionBinder;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Alteridem.Todo.Application;
-using Alteridem.Todo.Application.Commands.Add;
-using Alteridem.Todo.Application.Commands.Append;
-using Alteridem.Todo.Application.Commands.Archive;
-using Alteridem.Todo.Application.Commands.Delete;
-using Alteridem.Todo.Application.Commands.Deprioritize;
-using Alteridem.Todo.Application.Commands.Do;
-using Alteridem.Todo.Application.Commands.Prepend;
-using Alteridem.Todo.Application.Commands.Priority;
-using Alteridem.Todo.Application.Commands.Replace;
-using Alteridem.Todo.Application.Queries.IndividualTask;
-using Alteridem.Todo.Application.Queries.List;
-using Alteridem.Todo.Application.Queries.ListContexts;
-using Alteridem.Todo.Application.Queries.ListPriority;
-using Alteridem.Todo.Application.Queries.ListProjects;
+using Alteridem.Todo.Application.Commands;
+using Alteridem.Todo.Application.Queries;
 using Alteridem.Todo.Domain.Entities;
 using Alteridem.Todo.Domain.Interfaces;
 using Alteridem.Todo.Extensions;
