+++ tests/todo.tests/Application/Commands/DeleteTermCommandHandlerTests.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using Alteridem.Todo.Application.Commands.Delete;
+using Alteridem.Todo.Application.Commands;
 using Alteridem.Todo.Infrastructure.Persistence;
 using Alteridem.Todo.Tests.Mocks;
 using FluentAssertions;
