+++ tests/todo.tests/Application/Commands/DeprioritizeCommandHandlerTests.cs
@@ -4,8 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Alteridem.Todo.Application.Commands.Deprioritize;
-using Alteridem.Todo.Application.Commands.Priority;
+using Alteridem.Todo.Application.Commands;
 using Alteridem.Todo.Infrastructure.Persistence;
 using Alteridem.Todo.Tests.Mocks;
 using FluentAssertions;
