+++ tests/todo.tests/Application/Commands/DoTasksCommandHandlerTests.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Alteridem.Todo.Application.Commands.Do;
+using Alteridem.Todo.Application.Commands;
 using Alteridem.Todo.Infrastructure.Persistence;
 using Alteridem.Todo.Tests.Mocks;
 using FluentAssertions;
