+++ tests/todo.tests/Application/Queries/ListPriorityQueryHandlerTests.cs
@@ -2,9 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Alteridem.Todo.Application.Queries.List;
-using Alteridem.Todo.Application.Queries.ListPriority;
-using Alteridem.Todo.Domain.Common;
+using Alteridem.Todo.Application.Queries;
 using Alteridem.Todo.Domain.Interfaces;
 using Alteridem.Todo.Infrastructure.Persistence;
 using Alteridem.Todo.Tests.Mocks;
