+++ tests/todo.tests/Application/Queries/TaskQueryHandlerTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Alteridem.Todo.Application.Queries.IndividualTask;
+using Alteridem.Todo.Application.Queries;
 using Alteridem.Todo.Infrastructure.Persistence;
 using Alteridem.Todo.Tests.Mocks;
 using FluentAssertions;
